feat; separate copy pasted notification component
